feat(server): createFlowletFetchHandler — framework-agnostic routing; @flowlet/next becomes a thin adapter
