+++ packages/flowlet-next/src/handler.ts
@@ -10,210 +10,18 @@
  *   export const dynamic = "force-dynamic";
  *   export const { GET, POST } = createFlowletHandler();
  *
- * Endpoints (relative to the catch-all mount):
- *   POST /chat          — streamed agent turn (host tools via the caller seam)
- *   POST /action        — sandbox dispatch through the policy (+ approval tokens)
- *   GET|POST /integrations — Composio connect flow (inert without the key)
- *   GET  /capabilities  — { chat, integrations, voice } from env-key presence
- *   POST /tick          — drives the automations scheduler
- *
- * ZERO-CONFIG: with no options it reads `.env` (capability-additive keys) and
- * `.flowlet/` (theme, tools manifest) and just works — ANTHROPIC_API_KEY alone
- * gives working chat + generated UI.
+ * All routing/assembly lives in `@flowlet/server`'s framework-agnostic
+ * `createFlowletFetchHandler` — see its docblock for the endpoint list and
+ * the ZERO-CONFIG contract. This file just points both HTTP methods at it.
  */
-import type { ToolSet } from "ai";
-import { prewiredComponents } from "@flowlet/components/descriptors";
-import {
-  handleChat,
-  handleAction,
-  createApprovalStore,
-  DEFAULT_INTEGRATION_CATALOG,
-  createConnectionsStore,
-  handleIntegrationsGet,
-  handleIntegrationsPost,
-  detectCapabilities,
-  loadFlowletDir,
-  manifestToolsToHostTools,
-  buildInstructions,
-  createAgentCache,
-  createAutomationsWorld,
-  resolveModel,
-  type FlowletAutomationsWorld,
-  defaultFlowletPolicy,
-  resolvePrincipal,
-  DEFAULT_PRINCIPAL,
-  parseHandlerOptions,
-  type FlowletHandlerOptions,
-} from "@flowlet/server";
+import { createFlowletFetchHandler, type FlowletHandlerOptions } from "@flowlet/server";
 
 export interface FlowletRouteHandlers {
   GET: (req: Request) => Promise<Response>;
   POST: (req: Request) => Promise<Response>;
 }
 
-/** The path segment after `/api/flowlet/` the catch-all was invoked with. */
-function subPath(req: Request): string {
-  const segments = new URL(req.url).pathname.split("/").filter(Boolean);
-  return segments[segments.length - 1] ?? "";
-}
-
-export function createFlowletHandler(rawOptions: FlowletHandlerOptions = {}): FlowletRouteHandlers {
-  const options = parseHandlerOptions(rawOptions);
-
-  // Everything below is built lazily on first request: `next build` imports
-  // route modules at build time, and reading `.flowlet/` or requiring keys
-  // there would break builds in clean CI environments. `assemble` is async
-  // because provider resolution loads optional peer packages via dynamic
-  // import; the memoized promise keeps the lazy-on-first-request behavior.
-  let assembled: ReturnType<typeof assemble> | null = null;
-  async function assemble() {
-    const capabilities = detectCapabilities(undefined, {
-      hasInjectedModel: options.model !== undefined,
-    });
-    const loaded = loadFlowletDir(options.flowletDir);
-    const hostTools = options.hostTools ?? manifestToolsToHostTools(loaded.manifest.tools);
-    const model = options.model ?? (await resolveModel());
-    const policy = options.policy ?? defaultFlowletPolicy;
-    const catalog = options.integrations ?? DEFAULT_INTEGRATION_CATALOG;
-    const connections = options.connections ?? createConnectionsStore(catalog);
-
-    const world: FlowletAutomationsWorld | null =
-      options.automations === false
-        ? null
-        : createAutomationsWorld({
-            policy,
-            model,
-            ...(options.automations?.tools ? { tools: options.automations.tools } : {}),
-            scope: { tenantId: "flowlet-embedded", subject: DEFAULT_PRINCIPAL.userId },
-          });
-
-    const serverTools = (): ToolSet => {
-      const extra = typeof options.tools === "function" ? options.tools() : options.tools ?? {};
-      return { ...extra, ...(world ? world.authoringTools() : {}) };
-    };
-
-    const instructions =
-      options.instructions ??
-      buildInstructions({
-        productName: options.productName ?? "this app",
-        brand: loaded.brand,
-        components: options.components ?? [],
-        hostToolNames: hostTools.map((t) => t.name),
-        integrations: capabilities.integrations ? catalog : [],
-        automations: world !== null,
-        ...(options.instructionsExtra ? { extra: options.instructionsExtra } : {}),
-      });
-
-    const getAgent = createAgentCache({
-      model,
-      policy,
-      instructions,
-      // The engine's render_view registry must know the prewired catalog too —
-      // host-node validation rejects any name it can't find (ENG-186).
-      components: [...prewiredComponents, ...(options.components ?? [])],
-      tools: serverTools,
-      ...(capabilities.integrations ? { toolkits: () => connections.connectedToolkits() } : {}),
-      ...(options.cacheKey ? { cacheKey: options.cacheKey } : {}),
-      ...(options.maxSteps !== undefined ? { maxSteps: options.maxSteps } : {}),
-    });
-
-    return {
-      capabilities,
-      hostTools,
-      policy,
-      connections,
-      world,
-      serverTools,
-      getAgent,
-      approvals: createApprovalStore(),
-    };
-  }
-  // Memoize the assembly promise, but drop it on rejection so the next
-  // request retries fresh (the old sync code threw before assignment, so a
-  // boot failure — bad FLOWLET_MODEL, missing peer — was never cached). The
-  // identity check keeps a late rejection from clobbering a newer assembly.
-  const state = () => {
-    if (!assembled) {
-      const p = assemble();
-      p.catch(() => {
-        if (assembled === p) assembled = null;
-      });
-      assembled = p;
-    }
-    return assembled;
-  };
-
-  /** A boot (assembly) failure surfaces as a 500, not an unhandled rejection. */
-  function bootError(err: unknown): Response {
-    return Response.json(
-      { error: err instanceof Error ? err.message : String(err) },
-      { status: 500 },
-    );
-  }
-
-  async function GET(req: Request): Promise<Response> {
-    let s: Awaited<ReturnType<typeof assemble>>;
-    try {
-      s = await state();
-    } catch (err) {
-      return bootError(err);
-    }
-    switch (subPath(req)) {
-      case "capabilities":
-        return Response.json(s.capabilities);
-      case "integrations":
-        return handleIntegrationsGet(req, {
-          store: s.connections,
-          enabled: s.capabilities.integrations,
-          options,
-        });
-      default:
-        return Response.json({ error: "not found" }, { status: 404 });
-    }
-  }
-
-  async function POST(req: Request): Promise<Response> {
-    let s: Awaited<ReturnType<typeof assemble>>;
-    try {
-      s = await state();
-    } catch (err) {
-      return bootError(err);
-    }
-    switch (subPath(req)) {
-      case "chat":
-        return handleChat(req, {
-          getAgent: s.getAgent,
-          hostTools: s.hostTools,
-          options,
-          // capabilities.chat is the single source of truth: it already
-          // folds in an injected model (via hasInjectedModel above) alongside
-          // any configured provider key.
-          chatEnabled: s.capabilities.chat,
-        });
-      case "action":
-        return handleAction(req, {
-          getTools: s.serverTools,
-          policy: s.policy,
-          approvals: s.approvals,
-          options,
-        });
-      case "integrations":
-        return handleIntegrationsPost(req, {
-          store: s.connections,
-          enabled: s.capabilities.integrations,
-          options,
-        });
-      case "tick": {
-        if (!s.world) return Response.json({ error: "automations are disabled" }, { status: 404 });
-        const guard = await resolvePrincipal(req, options);
-        if (!guard.ok) return guard.response;
-        await s.world.tick();
-        return Response.json({ ok: true });
-      }
-      default:
-        return Response.json({ error: "not found" }, { status: 404 });
-    }
-  }
-
-  return { GET, POST };
+export function createFlowletHandler(options: FlowletHandlerOptions = {}): FlowletRouteHandlers {
+  const handler = createFlowletFetchHandler(options);
+  return { GET: handler, POST: handler };
 }
